refactor(org-admin): split tickets page under the line cap (#544)
Extract the event tickets admin page (947 lines) into cohesive units:
ConfirmPaymentDialog.svelte (the inline PWYC confirm-payment modal, pwyc
value bound back to the page), checkin-adapter.ts (the AdminTicketSchema ->
CheckInDialog shape adapter), and ticket-member-admin.svelte.ts (a runes
factory bundling the membership + blacklist state, mutations, and handlers).
Ticket queries, sort/filter URL round-trip (#452 order_by), and lifecycle
mutations stay in the page. Zero behavior change.
